Extract KeyboardObserver
